+++ src/main/java/com/gmail/val59000mc/scenarios/scenariolisteners/CutCleanListener.java
@@ -100,7 +100,7 @@ public class CutCleanListener extends ScenarioListener{
 				count *= 2;
 			}
 
-			drop = new ItemStack(oreType.get().getDrop(), count);
+			drop = oreType.get().getDrop().getStack(count);
 			UhcItems.spawnExtraXp(loc,xp);
 		}