+++ src/main/java/com/gmail/val59000mc/scenarios/scenariolisteners/TimberListener.java
@@ -1,14 +1,21 @@
 package com.gmail.val59000mc.scenarios.scenariolisteners;
 
-import com.gmail.val59000mc.scenarios.Option;
-import com.gmail.val59000mc.scenarios.ScenarioListener;
-import com.gmail.val59000mc.utils.UniversalMaterial;
+import java.util.Set;
+
+import org.bukkit.Location;
 import org.bukkit.Material;
+import org.bukkit.World;
 import org.bukkit.block.Block;
-import org.bukkit.block.BlockFace;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.block.BlockBreakEvent;
 import org.bukkit.inventory.ItemStack;
+import org.bukkit.util.Vector;
+
+import com.gmail.val59000mc.players.UhcPlayer;
+import com.gmail.val59000mc.scenarios.Option;
+import com.gmail.val59000mc.scenarios.ScenarioListener;
+import com.gmail.val59000mc.utils.FloodFill;
+import com.gmail.val59000mc.utils.UniversalMaterial;
 
 public class TimberListener extends ScenarioListener {
 
@@ -18,46 +25,37 @@ public class TimberListener extends ScenarioListener {
 	@Option(key = "drop-planks")
 	private boolean dropPlanks = false;
 
+	@Option(key = "log-break-limit")
+	private int logBreakLimit = 1000;
 
 	@EventHandler
 	public void onBlockBreak(BlockBreakEvent e) {
-		Block block = e.getBlock();
+		final Block block = e.getBlock();
+		final ItemStack tool = e.getPlayer().getItemInHand();
 
-		if (UniversalMaterial.isLog(block.getType())) {
-			int brokenLogs = breakTree(block, 2);
+		if (UniversalMaterial.isLog(block.getType()) && UniversalMaterial.isAxe(tool.getType())) {
+			final Set<Vector> treeLogs = FloodFill.floodFind26(block, b -> UniversalMaterial.isLog(b.getType()), logBreakLimit);
+
+			final World world = block.getWorld();
+			for (Vector v : treeLogs) {
+				breakLog(world.getBlockAt(v.getBlockX(), v.getBlockY(), v.getBlockZ()));
+			}
+
 			if (calculateAxeDamage) {
-				ItemStack tool = e.getPlayer().getItemInHand();
-				if (UniversalMaterial.isAxe(tool.getType())) {
-					tool.setDurability((short) (tool.getDurability() + brokenLogs));
-				}
+				UhcPlayer.damageItemInMainHand(e.getPlayer(), treeLogs.size());
 			}
 		}
 	}
 
-	private int breakTree(Block block, int i) {
-		int broken = 0;
-		if (UniversalMaterial.isLog(block.getType())){
-			if (dropPlanks){
-				block.setType(Material.AIR);
-				block.getWorld().dropItem(block.getLocation(), new ItemStack(UniversalMaterial.OAK_PLANKS.getType(), 4));
-			}else {
-				block.breakNaturally();
-			}
-			broken++;
-			i = 2;
-		}else {
-			i--;
+	private void breakLog(Block log) {
+		if (dropPlanks) {
+			log.setType(Material.AIR);
+			final Location itemDropLocation = log.getLocation().add(0.5, 0.5, 0.5);
+			final ItemStack itemDrop = new ItemStack(UniversalMaterial.OAK_PLANKS.getType(), 4);
+			log.getWorld().dropItem(itemDropLocation, itemDrop);
+		} else {
+			log.breakNaturally();
 		}
-		if (i > 0){
-			for (BlockFace face : BlockFace.values()) {
-				if (face.equals(BlockFace.DOWN) || face.equals(BlockFace.UP) || face.equals(BlockFace.NORTH) ||
-						face.equals(BlockFace.EAST) || face.equals(BlockFace.SOUTH) || face.equals(BlockFace.WEST)) {
-					broken += breakTree(block.getRelative(face), i);
-				}
-			}
-		}
-
-		return broken;
 	}
 
 }