+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Bug fixes
 
+- Fixed a bug where the `generate-sugar-cane` option would not work on Minecraft 1.12.2 and below
 - Fixed a bug in the Timber scenario where some diagonal branches would not be chopped down
 - Fixed a StackOverflowError that could happen when chopping very large trees in the Timber scenario
 - Added missing log and leaf types to the Timber and Fast Leaves Decay scenarios